seperated the app.js
