+++ src/constants.js
@@ -9,3 +9,11 @@ export const USER_INTERFACE_ID = 'user-interface';
 export const START_QUIZ_BUTTON_ID = 'start-quiz-button';
 export const ANSWERS_LIST_ID = 'answers-list';
 export const NEXT_QUESTION_BUTTON_ID = 'next-question-button';
+export const QUESTION_CONTAINER='question-container'
+export const QUESTION='question'
+export const ANSWER_BUTTONS_ELEMENT='answer-buttons'
+export const SCORE_ELEMENT='score'
+export const CONTINUE_BUTTON='continue-button'
+export const RESULT_CONTAINER='result-container'
+export const RESULT='result'
+export const RESTART_BUTTON='restart-button'
