+++ src/pages/endingPage.js
@@ -1,0 +1,4 @@
+
+// show restart button
+// only if else functions
+
