+++ src/views/endingViews.js
@@ -1,0 +1,8 @@
+
+export const createAnswerElement = (key, answerText) => {
+    const element = document.createElement('li');
+    element.innerHTML = String.raw`
+      ${key}: ${answerText};
+    `;
+    return element;
+  };
