+++ src/views/welcomeView.js
@@ -7,6 +7,7 @@ import { START_QUIZ_BUTTON_ID } from '../constants.js';
 export const createWelcomeElement = () => {
   const element = document.createElement('div');
   element.innerHTML = String.raw`
+  
     <h1>Welcome</h1>
     <button id="${START_QUIZ_BUTTON_ID}">start quiz</button>
   `;
